Routing
Routed everything correctly. I think. I'll have to recheck my diagramme, and do some rerouting, but this is the basic mods and uses that I needed to get things connected in a logical fashion.

Basically, everything that needs performed feeds into the main() function, whereas the rest of the input() function feeds into image_rec() and voice_rec() functions, which will eventually change things that are referenced in the output() function which is referenced in the main() function after the input() function.

This will need to be changed most likely, but this is the basic code structure I want. Obviously all of the functions will need to be referenced in main() so that they execute.

This is based on the code from my previous years in VRC, examples of which can be found in my vex-in-the-zone-2017 repository.